+++ MoreShipUpgrades/Managers/RandomizeUpgradeManager.cs
@@ -1,8 +1,6 @@
-﻿using MoreShipUpgrades.Misc.TerminalNodes;
+﻿using MoreShipUpgrades.UI.TerminalNodes;
 using MoreShipUpgrades.UpgradeComponents.TierUpgrades.Player;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace MoreShipUpgrades.Managers
 {