+++ MoreShipUpgrades/UI/Cursor/UpgradeCursorElement.cs
@@ -1,10 +1,11 @@
 ﻿using InteractiveTerminalAPI.UI.Cursor;
-using MoreShipUpgrades.Misc.TerminalNodes;
+using MoreShipUpgrades.Managers;
 using MoreShipUpgrades.Misc.Util;
+using MoreShipUpgrades.UI.TerminalNodes;
 using System.Text;
 using UnityEngine;
 
-namespace MoreShipUpgrades.Misc.UI.Cursor
+namespace MoreShipUpgrades.UI.Cursor
 {
     internal class UpgradeCursorElement : CursorElement
     {
@@ -14,7 +15,7 @@ namespace MoreShipUpgrades.Misc.UI.Cursor
         {
             StringBuilder sb = new StringBuilder();
             sb.Append(new string(LguConstants.WHITE_SPACE, 2));
-            string name = Node.Name.Length > LguConstants.NAME_LENGTH ? Node.Name.Substring(0, LguConstants.NAME_LENGTH) : Node.Name + new string(LguConstants.WHITE_SPACE, Mathf.Max(0, LguConstants.NAME_LENGTH-Node.Name.Length));
+            string name = Node.Name.Length > LguConstants.NAME_LENGTH ? Node.Name.Substring(0, LguConstants.NAME_LENGTH) : Node.Name + new string(LguConstants.WHITE_SPACE, Mathf.Max(0, LguConstants.NAME_LENGTH - Node.Name.Length));
             if (!Active(this))
             {
                 if (Node.Unlocked && Node.CurrentUpgrade >= Node.MaxUpgrade)
@@ -36,12 +37,8 @@ namespace MoreShipUpgrades.Misc.UI.Cursor
             sb.Append(LguConstants.WHITE_SPACE);
             if (remainingLevels > 0)
             {
-                sb.Append(Node.GetCurrentPrice() + "$");
-                if (Node.SalePercentage < 1f)
-                {
-                    sb.Append(LguConstants.WHITE_SPACE);
-                    sb.Append($"({(1 - Node.SalePercentage) * 100:F0}% OFF)");
-                }
+                AppendPriceText(ref sb);
+                AppendSaleText(ref sb);
             }
             else
             {
@@ -50,5 +47,32 @@ namespace MoreShipUpgrades.Misc.UI.Cursor
             if (!Active(this)) sb.Append(LguConstants.COLOR_FINAL_FORMAT);
             return sb.ToString();
         }
+
+        void AppendPriceText(ref StringBuilder sb)
+        {
+            int price = Node.GetCurrentPrice();
+            int currentCredits = UpgradeBus.Instance.GetTerminal().groupCredits;
+            if (price <= currentCredits)
+            {
+                sb.Append(price + "$");
+            }
+            else
+            {
+                sb.Append(string.Format(LguConstants.COLOR_INITIAL_FORMAT, LguConstants.HEXADECIMAL_DARK_RED));
+                sb.Append(price);
+                sb.Append(LguConstants.COLOR_FINAL_FORMAT);
+                sb.Append("$");
+            }
+        }
+
+        void AppendSaleText(ref StringBuilder sb)
+        {
+            if (Node.SalePercentage < 1f)
+            {
+                sb.Append(LguConstants.WHITE_SPACE);
+                sb.Append($"({(1 - Node.SalePercentage) * 100:F0}% OFF)");
+            }
+        }
+
     }
 }