+++ MoreShipUpgrades/UpgradeComponents/OneTimeUpgrades/Store/Sigurd.cs
@@ -1,13 +1,13 @@
 ﻿using MoreShipUpgrades.Managers;
 using MoreShipUpgrades.Misc;
-using MoreShipUpgrades.Misc.TerminalNodes;
 using MoreShipUpgrades.Misc.Upgrades;
+using MoreShipUpgrades.UI.TerminalNodes;
 using MoreShipUpgrades.UpgradeComponents.Interfaces;
 using UnityEngine;
 
 namespace MoreShipUpgrades.UpgradeComponents.OneTimeUpgrades.Store
 {
-    class Sigurd : OneTimeUpgrade, IUpgradeWorldBuilding
+    public class Sigurd : OneTimeUpgrade, IUpgradeWorldBuilding
     {
         public const string UPGRADE_NAME = "Sigurd Access";
         internal const string WORLD_BUILDING_TEXT = "\n\nSigurd always laughed at Desmond when he remembered the stories about The Company paying 120% of the value of the" +
@@ -26,15 +26,11 @@ namespace MoreShipUpgrades.UpgradeComponents.OneTimeUpgrades.Store
         {
             Instance = this;
         }
-        public static float GetBuyingRateLastDay(float defaultValue)
+
+        public enum FunctionModes
         {
-            LategameConfiguration config = GetConfiguration();
-            if (!config.SIGURD_LAST_DAY_ENABLED.Value) return defaultValue;
-            if (!GetActiveUpgrade(UPGRADE_NAME)) return defaultValue;
-            System.Random random = new(StartOfRound.Instance.randomMapSeed);
-            if (random.Next(0, 100) < Mathf.Clamp(config.SIGURD_LAST_DAY_CHANCE.Value, 0, 100))
-                return defaultValue + (config.SIGURD_LAST_DAY_PERCENT.Value / 100f);
-            return defaultValue;
+            LastDay,
+            AllDays,
         }
 
         public static float GetBuyingRate(float defaultValue)
@@ -42,7 +38,18 @@ namespace MoreShipUpgrades.UpgradeComponents.OneTimeUpgrades.Store
             LategameConfiguration config = GetConfiguration();
             if (!config.SIGURD_ENABLED.Value) return defaultValue;
             if (!GetActiveUpgrade(UPGRADE_NAME)) return defaultValue;
-            if (TimeOfDay.Instance.daysUntilDeadline == 0) return defaultValue;
+            switch(config.SIGURD_MODE.Value)
+            {
+                case FunctionModes.LastDay:
+                    {
+                        if (TimeOfDay.Instance.daysUntilDeadline != 0) return defaultValue;
+                        break;
+                    }
+                case FunctionModes.AllDays:
+                    {
+                        break;
+                    }
+            }
 
             System.Random random = new(StartOfRound.Instance.randomMapSeed);
             if (random.Next(0, 100) < Mathf.Clamp(config.SIGURD_CHANCE.Value, 0, 100))
@@ -60,7 +67,7 @@ namespace MoreShipUpgrades.UpgradeComponents.OneTimeUpgrades.Store
             return WORLD_BUILDING_TEXT;
         }
 
-        public override bool CanInitializeOnStart => (GetConfiguration().SIGURD_ENABLED.Value || GetConfiguration().SIGURD_LAST_DAY_ENABLED.Value) && GetConfiguration().SIGURD_PRICE.Value <= 0;
+        public override bool CanInitializeOnStart => (GetConfiguration().SIGURD_ENABLED.Value) && GetConfiguration().SIGURD_PRICE.Value <= 0;
         public new static (string, string[]) RegisterScrapToUpgrade()
         {
             return (UPGRADE_NAME, GetConfiguration().SIGURD_ACCESS_ITEM_PROGRESSION_ITEMS.Value.Split(","));
@@ -75,7 +82,7 @@ namespace MoreShipUpgrades.UpgradeComponents.OneTimeUpgrades.Store
 
             return UpgradeBus.Instance.SetupOneTimeTerminalNode(UPGRADE_NAME,
                                     shareStatus: true,
-                                    configuration.SIGURD_ENABLED.Value || configuration.SIGURD_LAST_DAY_ENABLED.Value,
+                                    configuration.SIGURD_ENABLED.Value,
                                     configuration.SIGURD_PRICE.Value,
                                     configuration.OVERRIDE_UPGRADE_NAMES ? configuration.SIGURD_ACCESS_OVERRIDE_NAME : "");
         }