+++ MoreShipUpgrades/UpgradeComponents/TierUpgrades/Enemies/Hunter.cs
@@ -1,8 +1,8 @@
 ﻿using MoreShipUpgrades.API;
 using MoreShipUpgrades.Managers;
 using MoreShipUpgrades.Misc;
-using MoreShipUpgrades.Misc.TerminalNodes;
 using MoreShipUpgrades.Misc.Upgrades;
+using MoreShipUpgrades.UI.TerminalNodes;
 using MoreShipUpgrades.UpgradeComponents.Interfaces;
 using System.Collections.Generic;
 using System.Linq;