+++ MoreShipUpgrades/UpgradeComponents/TierUpgrades/Player/NightVision.cs
@@ -11,9 +11,9 @@ using UnityEngine.InputSystem;
 using MoreShipUpgrades.Input;
 using UnityEngine.UI;
 using MoreShipUpgrades.Misc.Util;
-using MoreShipUpgrades.Misc.TerminalNodes;
 using MoreShipUpgrades.UpgradeComponents.Interfaces;
 using LethalLib;
+using MoreShipUpgrades.UI.TerminalNodes;
 
 namespace MoreShipUpgrades.UpgradeComponents.TierUpgrades.Player
 {